requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
boto3==1.37.37
boto3-stubs==1.37.37
botocore==1.37.37
botocore-stubs==1.37.37
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ecdsa==0.19.1
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mypy-boto3-cognito-idp==1.37.13.post1
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-jose==3.4.0
python-socketio==5.13.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.5
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
types-awscrt==0.26.1
types-s3transfer==0.11.5
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
